refactor(types): consolidate host monitoring types - Merge separate type files (cpu, disk, mem, system) into single host.types.ts - Update component imports to use new type structure - Standardize naming conventions across monitoring components - Fix file extension for mem.types (from .tsx to .ts) - Improve type organization for better maintainability
