+++ src/components/HostInfo.tsx
@@ -2,7 +2,7 @@ import { Box, Card, CardHeader, Typography, CardContent, CircularProgress } from
 import Grid from '@mui/material/Grid2';
 import { useEffect, useState, JSX } from 'react';
 import { api } from '../services/apiService';
-import { SystemInfo } from '../types/system.types';
+import { SystemInfo } from '../types/host.types';
 
 export default function HostInfo(): JSX.Element {
     // State Management
