--- src/types/cpu.types.ts
@@ -1,7 +1,0 @@
-export interface CpuData {
-    cpu: string;
-    total: number;
-    idle: number;
-    used: number;
-    usage: number;
-}
