--- src/types/disk.types.ts
@@ -1,8 +1,0 @@
-export interface DiskData {
-    Filesystem: string;
-    Size: string;
-    Used: string;
-    Avail: string;
-    Use: string;
-    Mounted: string;
-}
