--- src/types/mem.types.tsx
@@ -1,5 +1,0 @@
-export interface MemData {
-    total: string;
-    used: string;
-    available: string;
-}
