--- src/types/system.types.ts
@@ -1,8 +1,0 @@
-export interface SystemInfo {
-    Hostname: string;
-    KernelName: string;
-    KernelRelease: string;
-    OperatingSystemPrettyName: string;
-    HardwareVendor: string;
-    HardwareModel: string;
-}
